Add Pmw brew
The main problem with Pmw is that it is not currently in PyPI. There
is a question on the mailing list about asking them to add it to PyPI
(2009), but it received no response:

http://sourceforge.net/tracker/?func=detail&aid=2866553&group_id=10743&atid=360743

Additionally, the directory appears to be incompatible with how pip
currently works. Calling `pip install ${PMW_TARBALL_URL}` fails with an
IOError because it cannot find `setup.py`. From what I can see, this is
a problem in how Pmw is structured and I do not think it is a failure of
pip.